move ModuleSpecifier to core/modules.rs
