+++ cli/compiler.rs
@@ -247,8 +247,8 @@ mod tests {
   fn test_compile_sync() {
     tokio_util::init(|| {
       let specifier = "./tests/002_hello.ts";
-      use crate::worker;
-      let module_name = worker::ModuleSpecifier::resolve_root(specifier)
+      use deno::ModuleSpecifier;
+      let module_name = ModuleSpecifier::resolve_root(specifier)
         .unwrap()
         .to_string();
 
@@ -294,8 +294,8 @@ mod tests {
   #[test]
   fn test_bundle_async() {
     let specifier = "./tests/002_hello.ts";
-    use crate::worker;
-    let module_name = worker::ModuleSpecifier::resolve_root(specifier)
+    use deno::ModuleSpecifier;
+    let module_name = ModuleSpecifier::resolve_root(specifier)
       .unwrap()
       .to_string();
