+++ cli/state.rs
@@ -13,10 +13,10 @@ use crate::permissions::DenoPermissions;
 use crate::progress::Progress;
 use crate::resources;
 use crate::resources::ResourceId;
-use crate::worker::ModuleSpecifier;
 use crate::worker::Worker;
 use deno::Buf;
 use deno::Loader;
+use deno::ModuleSpecifier;
 use deno::Op;
 use deno::PinnedBuf;
 use futures::future::Either;
