+++ core/Cargo.toml
@@ -19,6 +19,7 @@ lazy_static = "1.3.0"
 libc = "0.2.55"
 log = "0.4.6"
 serde_json = "1.0.39"
+url = "1.7.2"
 
 [[example]]
 name = "deno_core_http_bench"
